\name{seacarb_test}
\docType{data}
\alias{seacarb_test}
\title{Test data file to practice the use of seacarb}
\description{
Variables:

\item flag indicating which couple of variables is used
\item pH
\item total alkalinity
\item salinity
\item temperature
\item hydrostatic pressure
\item k1 and k2 flag (constants of Roy et al.)
\item pH scale (total scale)
}
\usage{seacarb_test}
\format{A data frame with 15 rows and 8 variables}
\source{None, these data were invented for this purpose}
\references{}
\keyword{datasets}
